loader/dri3: Do not wait for source buffer for prefilling unless front.
The source buffer should have been just swapped off and should already
be okay to blit from without any further synchronization unless it is
the front buffer.

If the server has flipped the buffer instead of copied it, then the fence
is not triggered until the frame is idled after the next frame is displayed.
However, the next frame cannot be presented since the back buffer preparation
is blocked on the source fence and deadlock occurs.

Signed-off-by: David Riley <davidriley@chromium.org>